requirement.txt
logzero==1.5.0
mysql-connector-python==8.0.20
numpy==1.18.5
pandas==1.0.5
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
win32-setctime==1.0.1
